refactor: Voucher 인터페이스에서 추상 클래스로 수정
