+++ src/main/java/org/prgrms/springbootbasic/entity/voucher/Voucher.java
@@ -1,0 +1,13 @@
+package org.prgrms.springbootbasic.entity.voucher;
+
+import java.util.UUID;
+
+public abstract class Voucher {
+    protected final UUID voucherId = UUID.randomUUID();
+
+    public UUID getVoucherId() {
+        return voucherId;
+    }
+
+    public abstract long discount(long beforeDiscount);
+}
