+++ src/main/java/org/prgrms/springbootbasic/factory/VoucherFactory.java
@@ -1,7 +1,8 @@
 package org.prgrms.springbootbasic.factory;
 
-import org.prgrms.springbootbasic.voucher.Voucher;
+import org.prgrms.springbootbasic.entity.voucher.Voucher;
 
 public interface VoucherFactory {
-    Voucher createVoucher(long factor);
+
+    Voucher createVoucher(long quantity);
 }
