--- src/main/java/org/prgrms/springbootbasic/voucher/Voucher.java
@@ -1,9 +1,0 @@
-package org.prgrms.springbootbasic.voucher;
-
-
-import java.util.UUID;
-
-public interface Voucher {
-    UUID getVoucherId();
-    long discount(long beforeDiscount);
-}
